Vorbereitung für das View Modell was persistente Städtenamen als Parameter empfängt

+++ Wetter/Wetter/Models/Weather/Location/LocationList.swift
@@ -8,9 +8,8 @@
 import Foundation
 
 struct LocationList: Codable {
-    
     var name: String
-    var lat: Double
-    var lon: Double
+    var lat: Float
+    var lon: Float
     
 }
